// src/modules/storage/enums/media-category.enum.ts
export enum MediaCategory {
  AVATAR = 'avatar',
  ROOM_COVER = 'room_cover',
  ROOM_THUMBNAIL = 'room_thumbnail',
  ROOM_BG = 'room_bg',
  GIFT_ICON = 'gift_icon',
  GIFT_ANIMATION = 'gift_animation',
  GIFT_PREVIEW = 'gift_preview',
  ANNOUNCEMENT_BANNER = 'announcement_banner',
  ANNOUNCEMENT_THUMBNAIL = 'announcement_thumbnail',
  ANNOUNCEMENT_ATTACHMENT = 'announcement_attachment',
  HOST_ID = 'host_id',
  HOST_PHOTO = 'host_photo',
  HOST_DOCUMENT = 'host_document',
  AGENCY_LOGO = 'agency_logo',
  AGENCY_BANNER = 'agency_banner',
  CHAT_IMAGE = 'chat_image',
  CHAT_DOCUMENT = 'chat_document',
  CHAT_AUDIO = 'chat_audio',
  CHAT_ATTACHMENT = 'chat_attachment',
}